refactoring: use react query to fetch data, using mock data
