+++ package.json
@@ -13,6 +13,9 @@
     "@chakra-ui/react": "^2.8.2",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
+    "@tanstack/react-query": "^5.31.0",
+    "@tanstack/react-query-devtools": "^5.31.0",
+    "axios": "^1.6.8",
     "framer-motion": "^11.1.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
