+++ src/components/ProductCard.tsx
@@ -7,6 +7,7 @@ import {
   Text,
 } from "@chakra-ui/react";
 import { Product } from "../interfaces/product";
+import image from "../assets/toaster-image.jpg";
 
 interface Props {
   product: Product;
@@ -17,7 +18,7 @@ const ProductCard = ({ product }: Props) => {
     <Card>
       <CardHeader fontSize={24}>{product.Name}</CardHeader>
       <CardBody>
-        <Image src={product.ImageUrl} />
+        <Image src={image} />
       </CardBody>
       <CardFooter>
         <Text>{product.Price},-</Text>
