+++ src/interfaces/product.ts
@@ -1,5 +1,4 @@
 export interface Product {
     Name: string,
-    Price: number,
-    ImageUrl: string
+    Price: number
 }
