+++ src/main.tsx
@@ -2,11 +2,16 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import { ChakraProvider } from "@chakra-ui/react";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <ChakraProvider>
-      <App />
+      <QueryClientProvider client={new QueryClient()}>
+        <App />
+        <ReactQueryDevtools />
+      </QueryClientProvider>
     </ChakraProvider>
   </React.StrictMode>
 );
